+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added the ability to show images in items and spells on a page with the `show-image` property.
+- Added support for automatically creating Encounter+'s new Roll Tables when exporting to module.
 
 ### Changed
 - Updated dependencies and webpack options for security and performance improvements.